@@ -12,13 +12,13 @@
     "@fortawesome/fontawesome-free": "^5.15.1",
     "bootstrap": "^4.5.3",
     "font-awesome": "^4.7.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^5.1.2",
-    "react-select": "^3.1.0",
-    "reactstrap": "^5.0.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "2.3.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-redux": "9.1.2",
+    "react-select": "5.8.3",
+    "reactstrap": "9.2.3",
+    "redux": "5.0.1",
+    "redux-thunk": "3.1.0",
     "simple-line-icons": "^2.5.5"
   }
 }
