@@ -39,26 +39,26 @@
     "lodash.uniq": "^4.5.0",
     "moment": "^2.29.1",
     "prettier": "^2.3.2",
-    "react": "^16.13.1",
-    "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "^16.13.1",
-    "react-html-parser": "^2.0.2",
-    "react-json-tree": "^0.13.0",
+    "react": "18.3.1",
+    "react-copy-to-clipboard": "5.1.0",
+    "react-dom": "18.3.1",
+    "html-react-parser": "^5.1.18",
+    "react-json-tree": "0.19.0",
     "react-jsonschema-form-bs4": "^1.7.1",
-    "react-redux": "^5.1.2",
-    "react-router-dom": "^4.3.1",
-    "react-select": "^3.1.0",
-    "reactstrap": "^5.0.0",
+    "react-redux": "9.1.2",
+    "react-router-dom": "7.0.1",
+    "react-select": "5.8.3",
+    "reactstrap": "9.2.3",
     "reconnecting-websocket": "^4.4.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "2.3.0",
+    "redux": "5.0.1",
+    "redux-thunk": "3.1.0",
     "sass": "^1.81.0",
     "sass-loader": "^16.0.3",
     "simple-line-icons": "^2.5.5",
     "style-loader": "^2.0.0",
-    "webpack": "^5.96.1",
-    "webpack-bundle-analyzer": "^3.9.0",
-    "webpack-cli": "^4.2.0"
+    "webpack": "5.96.1",
+    "webpack-bundle-analyzer": "4.10.2",
+    "webpack-cli": "5.1.4"
   },
   "scripts": {
     "prepublishOnly": "npm run clean && npm run build",
