@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { HashRouter, Route, Switch } from 'react-router-dom'
+import { HashRouter, Route, Routes } from 'react-router-dom'
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware } from 'redux'
-import thunk from 'redux-thunk'
+import {thunk} from 'redux-thunk'
 
 // Styles
 // Import Font Awesome Icons Set
@@ -303,11 +303,11 @@ __webpack_init_sharing__('default')
 // eslint-disable-next-line react/no-deprecated
 ReactDOM.render(
   <HashRouter>
-    <Switch>
+    <Routes>
       <Provider store={store}>
         <Route path="/" name="Home" component={Full} />
       </Provider>
-    </Switch>
+    </Routes>
   </HashRouter>,
   document.getElementById('root')
 )
