@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import JSONTree from 'react-json-tree'
+import {JSONTree} from 'react-json-tree'
 import {
   Card,
   CardHeader,
